rename: ScreenType 클래스 이름을 AdminType으로 변경
