+++ packy-domain/src/main/java/com/dilly/admin/adaptor/AdminGiftBoxReader.java
@@ -1,8 +1,8 @@
 package com.dilly.admin.adaptor;
 
-import com.dilly.admin.domain.giftbox.ScreenType;
 import com.dilly.gift.dao.AdminGiftBoxRepository;
-import com.dilly.gift.domain.giftbox.AdminGiftBox;
+import com.dilly.gift.domain.giftbox.admin.AdminGiftBox;
+import com.dilly.gift.domain.giftbox.admin.AdminType;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 
@@ -12,7 +12,7 @@ public class AdminGiftBoxReader {
 
     private final AdminGiftBoxRepository adminGiftBoxRepository;
 
-    public AdminGiftBox findByScreenType(ScreenType screenType) {
-        return adminGiftBoxRepository.findByScreenType(screenType);
+    public AdminGiftBox findByAdminType(AdminType adminType) {
+        return adminGiftBoxRepository.findByAdminType(adminType);
     }
 }
