--- packy-domain/src/main/java/com/dilly/admin/domain/giftbox/ScreenType.java
@@ -1,5 +1,0 @@
-package com.dilly.admin.domain.giftbox;
-
-public enum ScreenType {
-    ONBOARDING
-}
