+++ packy-domain/src/main/java/com/dilly/gift/dao/AdminGiftBoxRepository.java
@@ -1,10 +1,10 @@
 package com.dilly.gift.dao;
 
-import com.dilly.admin.domain.giftbox.ScreenType;
-import com.dilly.gift.domain.giftbox.AdminGiftBox;
+import com.dilly.gift.domain.giftbox.admin.AdminGiftBox;
+import com.dilly.gift.domain.giftbox.admin.AdminType;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface AdminGiftBoxRepository extends JpaRepository<AdminGiftBox, Long> {
 
-    AdminGiftBox findByScreenType(ScreenType screenType);
+    AdminGiftBox findByAdminType(AdminType adminType);
 }
