+++ packy-domain/src/main/java/com/dilly/gift/domain/giftbox/admin/AdminType.java
@@ -1,0 +1,5 @@
+package com.dilly.gift.domain.giftbox.admin;
+
+public enum AdminType {
+    ONBOARDING
+}
